Move API client to new module
